Add RSS feed
I started with the Jokes tutorial RSS sample code from the Remix
tutorial and then ran it through the validator and found a couple
issues. I ended up skimming through the spec and it turns out not to be
that long. I would recommend reading the docs for each element included.

See the RSS spec at https://www.rssboard.org/rss-specification

I'm using the path `/rss.xml` because that's what I used before, and
there's no other feeds yet on my site, so there's not a strong reason to
move it. In the future if there are multiple feeds, I might relocate
this to `/writing.rss.xml` to be more semantic.

Also some tweaks to the copy text on other pages.